refactor: model 폴더에 있던 DTO들 위치, 이름 정상화
