+++ app/src/main/java/com/wafflestudio/snutt2/data/mapper/SemesterStatusMappers.kt
@@ -1,0 +1,9 @@
+package com.wafflestudio.snutt2.data.mapper
+
+import com.wafflestudio.snutt2.domainmodel.SemesterStatus
+import com.wafflestudio.snutt2.lib.network.dto.SemesterStatusDto
+
+fun SemesterStatusDto.toDomain(): SemesterStatus = SemesterStatus(
+    current = current?.toDomain(),
+    next = next.toDomain(),
+)
