+++ app/src/main/java/com/wafflestudio/snutt2/data/semester_status/SemesterStatusRepositoryImpl.kt
@@ -5,8 +5,8 @@ import com.wafflestudio.snutt2.domainmodel.SemesterStatus
 import com.wafflestudio.snutt2.lib.network.Result
 import com.wafflestudio.snutt2.lib.network.SNUTTRestApi
 import com.wafflestudio.snutt2.lib.network.toDomainError
+import com.wafflestudio.snutt2.data.mapper.toDomain
 import com.wafflestudio.snutt2.lib.toOptional
-import com.wafflestudio.snutt2.model.toDomainModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -20,14 +20,14 @@ class SemesterStatusRepositoryImpl @Inject constructor(
 ) : SemesterStatusRepository {
 
     private val _semesterStatus: MutableStateFlow<SemesterStatus?> =
-        MutableStateFlow(storage.semesterStatus.get().value?.toDomainModel())
+        MutableStateFlow(storage.semesterStatus.get().value?.toDomain())
     override val semesterStatus: StateFlow<SemesterStatus?> = _semesterStatus.asStateFlow()
 
     override suspend fun fetchSemesterStatus(): Result<Unit> {
         try {
             val response = api._getSemesterStatus()
             storage.semesterStatus.update(response.toOptional())
-            _semesterStatus.value = response.toDomainModel()
+            _semesterStatus.value = response.toDomain()
             return Result.Success(Unit)
         } catch (e: Exception) {
             return Result.Fail(e.toDomainError())
