+++ app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/GetSemesterStatusResult.kt
@@ -1,4 +1,4 @@
 package com.wafflestudio.snutt2.lib.network.dto
 
-import com.wafflestudio.snutt2.model.SemesterStatusDto
+import com.wafflestudio.snutt2.lib.network.dto.SemesterStatusDto
 typealias GetSemesterStatusResult = SemesterStatusDto
