+++ app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/LectureBuildingDto.kt
@@ -2,8 +2,7 @@ package com.wafflestudio.snutt2.lib.network.dto
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
-import com.wafflestudio.snutt2.model.Campus
-import com.wafflestudio.snutt2.model.GeoCoordinate
+import com.wafflestudio.snutt2.domainmodel.Campus
 
 @JsonClass(generateAdapter = true)
 data class LectureBuildingDto(
@@ -11,7 +10,7 @@ data class LectureBuildingDto(
     @param:Json(name = "buildingNumber") val buildingNumber: String,
     @param:Json(name = "buildingNameKor") val buildingNameKor: String? = null,
     @param:Json(name = "buildingNameEng") val buildingNameEng: String? = null,
-    @param:Json(name = "locationInDMS") val locationInDMS: GeoCoordinate,
-    @param:Json(name = "locationInDecimal") val locationInDecimal: GeoCoordinate,
+    @param:Json(name = "locationInDMS") val locationInDMS: GeoCoordinateDTO,
+    @param:Json(name = "locationInDecimal") val locationInDecimal: GeoCoordinateDTO,
     @param:Json(name = "campus") val campus: Campus,
 )
