+++ app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/SemesterStatusDto.kt
@@ -1,0 +1,16 @@
+package com.wafflestudio.snutt2.lib.network.dto
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class SemesterStatusDto(
+    val current: CourseBookDto?,
+    val next: CourseBookDto,
+) {
+    companion object {
+        val Default = SemesterStatusDto(
+            current = null,
+            next = CourseBookDto(semester = 0L, year = 0L),
+        )
+    }
+}
