--- app/src/main/java/com/wafflestudio/snutt2/model/Campus.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.snutt2.model
-
-enum class Campus {
-    GWANAK,
-    YEONGEON,
-    PYEONGCHANG,
-}
